Enforce detail access constraints in contact operations by default
Previously, Irremovable constraints were enforced but ReadOnly
constraints were not, for in-memory contact detail operations.

While it is simple to work around any constraints simply by recreating
an identical contact (including id) in-memory, and then saving the
modified detail into that contact, we should enforce the constraints
consistently, and provide API to allow the user to ignore the
constraints for the in-memory operations if they desire (thus leaving
enforcement up to the backend).